Various fixes, mainly in the DPLL algorithm
Fixed a match error in the congruence closure algorithm.

Main fix in DPLL algorithm is in conflictAnalysis, when searching for the UIP.
confl is allowed to be null, when c <= 1. status reset after backtracking does
not override a possible conflict found when enqueueing the decision literal in
the opposite polarity any more.