Fixed some problems in various parts of the code
In the middle of debugging. Some benchmarks which should be sat are unsat.
Using a second congruence closure object as a sanity checker after the dpll
engine was run.
Added useList backtracking. Fixed a bug where a wrong timestamp was set to
invalid when backtracking. Clearing pending queue in backtrack.