Setting of conflict/reason clause after T-conflict
After a theory conflict:
Switched back to using the UIP analysis code (with setting the conflict clause
correctly beforehand now). The UIP is correctly computed for the current
decision level, but for every other level in the implication graph the literals
are added immediately. I think that we could compute a UIP for each decision
level to make the learnt clause more succinct.